+++ packages/modules/narduk-mapkit/examples/tile-overlay-crossfade.ts
@@ -3,7 +3,7 @@ import {
   createMapKitTileOverlay,
   crossfadeMapKitOverlayOpacity,
   initializeMapKit,
-} from '@loganrenz/narduk-mapkit/client'
+} from '@narduk-geo/narduk-mapkit/client'
 
 interface MapKitNamespace {
   Coordinate: new (latitude: number, longitude: number) => unknown